Separate type relation logic from checker.ts
